refactor(model-zoom): move methods zoom bar specific to model-zoom
